ComputerStorePOS/src/test/java/com/aiden/computerstorepos/test/MonitorTest.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */

package com.aiden.computerstorepos.test;

import com.aiden.computerstorepos.conf.AppConfig;
import com.aiden.computerstorepos.domain.Monitor;
import com.aiden.computerstorepos.services.MonitorService;
import org.springframework.context.ApplicationContext;
import org.springframework.context.annotation.AnnotationConfigApplicationContext;
import org.testng.Assert;
import static org.testng.Assert.*;
import org.testng.annotations.AfterClass;
import org.testng.annotations.AfterMethod;
import org.testng.annotations.BeforeClass;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Test;

/**
 *
 * @author Aidem
 */
public class MonitorTest {
    private MonitorService service;
    public MonitorTest() {
    }

    @Test
    public void testMonitor() throws Exception {
        Monitor data = service.getMonitor();
        Assert.assertEquals(data.getProductNumber(),"23MP55HQ-P");
    }

    @BeforeClass
    public static void setUpClass() throws Exception {
    }

    @AfterClass
    public static void tearDownClass() throws Exception {
    }

    @BeforeMethod
    public void setUpMethod() throws Exception {
        ApplicationContext ctx = new AnnotationConfigApplicationContext(AppConfig.class);
        service = (MonitorService)ctx.getBean("Monitor");
    }

    @AfterMethod
    public void tearDownMethod() throws Exception {
    }
}
